+++ aws-lambda/lib/extended-nodejs-function.ts
@@ -3,7 +3,12 @@ import {FunctionAlarms, FunctionAlarmsOptions} from './function-alarms';
 import {FunctionDeployment} from './function-deployment';
 import {DeployedFunctionOptions} from './extended-function';
 import {RetentionDays} from 'aws-cdk-lib/aws-logs';
-import {Architecture, Runtime} from 'aws-cdk-lib/aws-lambda';
+import {
+  Architecture,
+  LoggingFormat,
+  Runtime,
+  SystemLogLevel,
+} from 'aws-cdk-lib/aws-lambda';
 import {Duration} from 'aws-cdk-lib';
 import * as fs from 'fs';
 import * as path from 'path';
@@ -58,6 +63,8 @@ export class ExtendedNodejsFunction extends NodejsFunction {
       runtime: Runtime.NODEJS_20_X, // change default from NODEJS_20_X
       depsLockFilePath: ExtendedNodejsFunction.findDepsLockFile(props.entry),
       ...props,
+      loggingFormat: props.loggingFormat ?? LoggingFormat.JSON,
+      systemLogLevelV2: SystemLogLevel.INFO,
       environment: {
         NODE_OPTIONS: '--enable-source-maps',
         ...props.environment,