+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Added exportAndOutputParameter to ExtendedStack to allow for exporting and outputting parameters in the stack with one call
+-   Default logging for all extended lambda functions is now JSON instead of TEXT
 
 ### Removed